requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
Flask==2.3.3
fonttools==4.42.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.3.7
